[COOK-2527] name all private recipes with a leading underscore
The emerging Chef community idiom is to name all "private" recipes with
a leading underscore. End users should only include one of the public
recipes in their run lists: server, node, proxy. In the context of this
cookbook the default recipe is also treated as no-op now.

We will also add the following new attributes

* `node['jenkins']['node']['agent_type']` - type of agent used to
  connect to slave nodes.
* `node['jenkins']['http_proxy`]['variant']` - variant/flavor of HTTP
  proxy to use.