Only download the remote file if it doesn't already exist
This may seem a bit weird, but instead of adding a `remote_file` reosurce
to the resource collection on each run, we are using a simple File.exists?
check to guard the execution. Since the executor is _heavily_ used, the
`remote_file` significantly clogs up the debugging output and isn't an ideal
user experience.

This behavior is exactly the same as `:create_if_missing`, but the resource
is not added to the collection if it already exists.