Rename server -> master
After internal conversations, the term "server" is misleading. Even though
they are called "servers" in Jenkins land, that terminology is confusing to
users. Just like "slaves" are actually called "nodes" in Jenkins-land.

In order to make this cookbook as friendly as possible, we have chosen to
adopt the master-slave terminology to broadly refer to the server-node Jenkins
internally exhibits.

Since "node" and "server" already have meaning in the Chef context, we found
users were confused by these terminologies. Additionally, the community
commonly refers to Jenkins masters and Jenkins slaves/builders. As such, we
chose to adopt this naming schema to reduce/eliminate confusion.