Accept 403 responses while waiting until the server is ready
After enabling authentication, any endpoint the CLI tries to hit will require
HTTP Basic Authentication. We do not want to provide a Basic Authentication
mechanism, since we don't use the HTTP API anywhere else.

If the server responds with an HTTP 403 (Unauthorized), we can safely assume
that Jenkins is running. The only way we would get that error is if Jenkins
is fully booted, so getting a 403 response in this case satisfies the
constraint that the server is "ready".

- Fixes #126
- Closes #127