Fix a bug in default windows domain name
Domain name default value is set to nil: when the cookbook generates
winsw xml file template domainname is set to "".

Unfortunately winsw ignore the whole  <serviceaccount> part of the
config when domain name is not specified or empty.

In the end, when domain is not specified in jenkins slave, user is
ignored and jenkins service will run as administrator..