Use atomic updates when downloading the slave JAR
On Windows, when slave.jar changes on the master, the 'slave_jar_resource'
remote_file will try to replace the file. This means we get

    Errno::EACCES: remote_file[C:\chef\cache/slave.jar] (dynamically defined) had an error: Errno::EACCES: Permission denied

error since the file is currently in use (java -jar slave.jar is running).
Setting atomic_update to false means we will just use cp, which works.