Replace service notifications with runit_service.
At the 1.0 release, the runit cookbook replaced it's definition with a resource and provider that extend the built-in chef service resource. However, it provided a wrapper for users, that created a generic service resource. This generic resource also ignored settings passed (like the timeout setting in runit). This commit replaces all of the instances of `notifies :foo, 'service[jenkins]'` with `notifies :foo, 'runit_service[jenkins]'`.