Recover from commands executed with unknown credentials
We'll fall back to executing the command without authentication if the
command fails in a very specific way. This is a sign the provided
private key is unknown to the Jenkins master. This exception is commonly
thrown the first time a Chef run enables authentication on the Jenkins
master. This should also fix some cases of JENKINS-22346.