Fetch credentials via ID instead of username
Looking up credentials by username means that you can't have a password
and a private key with the same username.  This leads to LOTS of
confusion (see #447).

This commit switches the helper to lookup existing credentials via ID
instead of by username so that you can have different cred typess with
the same username and also allows you to change a username on a
resource.

Since we now looking by the ID all the time, the conditionals inside of
ssh_slave are no longer needed.

Fixes #447

Signed-off-by: Evan Borgstrom <evan@borgstrom.ca>
Signed-off-by: Jennifer Davis <iennae@gmail.com>

Minor updates to the READMEfor PR #501