add use_inline_resources and use action DSL helper
upgrade to best-practices, as should be enforced by several
foodcritic rules.  everything seems to ultimately inherit from
LWRPBase in this cookbook, which is somewhat pointless and/or buggy
if you don't declare use_inline_resources and then use the action
DSL helper instead of using `def action` directly.

it seems this cookbook mostly constructs its own sub-resources manually
and does not place them in a resource collection and does not gather
their updated_by_last_action state, so notifications off of the outer
resources here are likely still broken (or maybe this is by design
and nobody wants notifies to fire off of the subresources in some
of these providers, but that seems doubtful?)

anyway, i don't think this is a behavior change, but starts to drag
the cookbook style into 2014/2015 at least.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>