remove super from actions and private
common code shouldn't use inheritance it should use mixins or
ideally composition -- the slave_jnlp providers should probably
use jenkins_slave sub-providers to do their work, but they do
not.

anyway super is busted, so do the minimum amount to move that to a
function which can then at least be called directly.

also it seems someone mistakenly thought they needed to use
`protected` so that subclasses could use methods, which is incorrect,
this is ruby, not java, subclasses can call all private methods in
the superclass (`protected` is weird in ruby -- it allows a receiver
to be a different instance of the same clas, which is occasionally
useful, but 99% of the time in our code `protected` is used when
`private` is meant).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>