Wait for Jenkins in case of EADDRNOTAVAIL
wait_until_ready! retries connecting to the Jenkins port in case of many
different exceptions being returned from the open() call.

In my case, I'm currently hitting Errno::EADDRNOTAVAIL for the time when
Jetty is down and _not_ returning 503s. This exception will currently
propagated to the caller and provisining fails.

Thus add Errno::EADDRNOTAVAIL to the list of expected exceptions when
connecting to Jenkins.

Signed-off-by: Steffen Gebert <steffen.gebert@typo3.org>