Remove usage of update-center.json for installing/updating plugins
When reconverging an airgapped machine with Jenkins, it attempts to load the update-center.json to get a list of plugin metadata to determine what needs to me updated.  This caused the converge to fail.  With this I was able to convege multiple times in a row and also update plugins if I downloaded and provided the copy of the plugin locally.