Forcibly upgrade config file.
This cookbook maintains the config file `/etc/default/jenkins` on
Debian and derivatives. If upstream maintainers change this file
in a new package version, apt will fail to install it and error
out.

The change here forces the cookbook to use the local copy of
`/etc/default/jenkins`, as it is under management and will be
rewritten on every converge anyway.

Signed-off-by: Don Luchini <don-code@users.noreply.github.com>