Resolve https://github.com/sous-chefs/jenkins/issues/759 by removing the 'touch' command executed as root (#760)
* Resolve https://github.com/sous-chefs/jenkins/issues/759 by removing the 'touch' command executed as root

The WAR-based installation does a 'touch' of a file in the Jenkins
home whilst running as root. A user that has sufficient access to
Jenkins (E.g. can run a shell script from a Jenkins job) can create a
symbolic link from /var/lib/jenkins/jenkins.start to any location on
the filesystem the next time the Jenkins service is restarted / server
rebooted.

E.g. sh("rm -f /var/lib/jenkins/jenkins.start; ln -s /this-file-ends-up-created-in-root-directory /var/lib/jenkins/jenkins.start");

The risk is relatively low, but it's bound to be possible to exploit
to produce a local denial of service.

The history of this command seems to date back to this cookbook making
use of the file to drive a Jenkins restart.  That code was removed
from recipes/server.rb in commit
f9fbc2fbd0eb2c25edf4254cc78bc7cd95d455ad, making the touch command
redundant but left in-place.

* Log change: Remove touch command run as root from .war-based service definition - [@davidsainty](https://github.com/davidsainty)

Co-authored-by: Jason Field <4923914+Xorima@users.noreply.github.com>