Specify credential by id.
To be honest I'm not sure what's _supposed_ to happen when you save the
result of a resource declaration to a variable but the credentials field
normally expects a credential id.

Without this change it seems like the resource name is converted to a
string and finds its way into a groovy script where it is invalid.