Pin the Jenkins version to the last compatible version.
This cookbook expects the package install of Jenkins to install with
sysvinit style init scripts and configures jenkins with that assumption.

Jenkins packages after 2.335 (weekly) and 2.332.1 (LTS) and on use
systemd service units to manage the Jenkins service [1].

This cookbook has not yet been made compatible with these versions.

This pull request is something of a straw man to get CI turning over
again so I can start working on a new cookbook version that supports the
current Jenkins packages.

[1]: https://www.jenkins.io/blog/2022/03/25/systemd-migration/