Update Jenkins repository key URLs.
The repository signing keys for the Jenkins apt and rpm repositories
have been rotated and the new signing keys are available at a separate
URL from the previous ones.

As of 2023-06-19 the previous keys are no longer available and the URLs
used prior to this patch 404 so I do not think there is an advantage to
making this change conditional.

Background on the Jenkins repository change is here: https://www.jenkins.io/blog/2023/03/27/repository-signing-keys-changing/